# verification/curator-service/api/openapi/openapi.yaml
openapi: 3.0.3
info:
    title: Global.health curator service API
    description: Server for CRUD operations on Global.health data
    termsOfService: 'https://global.health/terms-of-use/'
    contact:
        name: Global.health
        url: https://global.health
        email: info@global.health
    license:
        name: MIT
        url: 'https://opensource.org/licenses/MIT'
    version: 1.0.0
...
paths:
...
    /sources:
        get:
            tags: [Source]
            summary: Lists sources
            operationId: listSources
            parameters:
                - name: page
                  in: query
                  description: The pages of sources to skip before starting to collect the result set
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      default: 1
                - name: limit
                  in: query
                  description: The number of sources to return
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 10
                - name: url
                  in: query
                  description: Origin URL substring on which to filter the result set
                  required: false
                  allowEmptyValue: true
                  allowReserved: true
                  schema:
                      type: string
                - name: access_token
                  description: Bearer token used by automated ingestion lambdas.
                  in: query
                  required: false
                  schema:
                      type: string
            responses:
                '200':
                    $ref: '#/components/responses/200SourceArray'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        post:
            tags: [Source]
            summary: Creates a new source
            operationId: createSource
            requestBody:
                description: Source to add
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NewSource'
            responses:
                '201':
                    $ref: '#/components/responses/201Source'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
    /acknowledgment-sources:
        get:
            tags: [Source]
            summary: Lists sources for acknowledgments table
            operationId: listSourcesForTable
            responses:
                '200':
                    $ref: '#/components/responses/200AcknowledgmentSource'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
...
    /sources/{id}/retrieve:
        parameters:
            - name: id
              in: path
              description: The source ID for which to retrieve content
              required: true
              schema:
                  type: string
                  pattern: '^[a-f\d]{24}$'
            - name: parse_start_date
              in: query
              description: >
                  The date on which to begin parsing/ingesting data (inclusive).
                  Provided in YYYY-MM-DD format.
              required: false
              schema:
                  type: string
                  pattern: '^[\d]{4}-[\d]{2}-[\d]{2}$'
            - name: parse_end_date
              in: query
              description: >
                  The date on which to end parsing/ingesting data (inclusive).
                  Provided in YYYY-MM-DD format.
              required: false
              schema:
                  type: string
                  pattern: '^[\d]{4}-[\d]{2}-[\d]{2}$'
        post:
            tags: [Source]
            summary: Trigger a retrieval of the content for the specified source
            operationId: retrieve
            responses:
                '200':
                    $ref: '#/components/responses/200Retrieval'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '404':
                    $ref: '#/components/responses/404'
                '500':
                    $ref: '#/components/responses/500'
    /sources/uploads:
        get:
            tags: [Source, Upload]
            summary: Lists all uploads for all sources
            operationId: listUploads
            parameters:
                - name: page
                  in: query
                  description: The pages of uploads to skip before starting to collect the result set
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      default: 1
                - name: limit
                  in: query
                  description: The number of uploads to return
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 10
                - name: changes_only
                  in: query
                  description: >
                      Filters successful uploads to only include those that created or modified 
                      cases. The response will still include non-successful uploads.
                  required: false
                  schema:
                      type: boolean
                      default: false
            responses:
                '200':
                    $ref: '#/components/responses/200UploadsArray'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '500':
                    $ref: '#/components/responses/500'
...
    /sources/parsers:
        get:
            summary: Lists parsers available for automated ingestion
            tags: [Suggest]
            operationId: listParsers
            responses:
                '200':
                    $ref: '#/components/responses/200ParsersArray'
                '500':
                    $ref: '#/components/responses/500'
    /cases:
        get:
            tags: [Case]
            summary: Lists cases
            operationId: listCases
            parameters:
                - name: page
                  in: query
                  description: The pages of sources to skip before starting to collect the result set
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      default: 1
                - name: limit
                  in: query
                  description: The number of cases to return
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 10
                - name: count_limit
                  in: query
                  description: The maximum number of documents that will be counted in mongoDB to make queries faster
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 100
                      default: 10000
                      maximum: 50000
                - name: sort_by
                  in: query
                  description: Keyword to sort by
                  required: false
                  schema:
                      type: string
                - name: order
                  in: query
                  description: Sorting order
                  required: false
                  schema:
                      type: string
                - name: q
                  in: query
                  description: The search query
                  required: false
                  allowEmptyValue: true
                  allowReserved: true
                  schema:
                      type: string
            responses:
                '200':
                    $ref: '#/components/responses/200CaseArray'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        post:
            tags: [Case]
            summary: Creates one (or multiple identical) new cases
            operationId: createCase
            requestBody:
                description: Case to add
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NewCase'
            parameters:
                - name: validate_only
                  in: query
                  description: Whether to validate the case without creating it
                  required: false
                  schema:
                      type: boolean
                      default: false
                  allowEmptyValue: true
                - name: num_cases
                  in: query
                  description: The number of cases to create
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      default: 1
            responses:
                '201':
                    $ref: '#/components/responses/201Create'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        put:
            tags: [Case]
            summary: Upserts a case
            description: >
                Creates or updates a case, depending on whether the provided values for
                the `caseReference.sourceId` and `caseReference.sourceEntryId` fields
                correspond to an existing case record.
            operationId: upsertCase
            requestBody:
                description: Case to upsert
                required: true
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: '#/components/schemas/NewCase'
                            required:
                                - caseReference
            responses:
                '200':
                    $ref: '#/components/responses/200Case'
                '201':
                    $ref: '#/components/responses/201Case'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        delete:
            summary: >
                Deletes multiple cases. It is required to supply exactly one of either 
                caseIds or query in the request body. If caseIds are supplied, cases 
                corresponding to those caseIds will be deleted. If query is supplied, 
                all cases that match the query will be deleted.
            tags: [Case]
            operationId: deleteCases
            requestBody:
                description: Cases to delete
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                caseIds:
                                    description: Cases corresponding to these ids will be deleted
                                    type: array
                                    items:
                                        type: integer
                                        format: int32
                                        minimum: 1
                                query:
                                    description: >
                                        Cases matching this query will be deleted. Must contain
                                        non-whitespace characters.
                                    type: string
                                    pattern: \S+
                            oneOf:
                                - required: [caseIds]
                                - required: [query]
            responses:
                '204':
                    description: Cases deleted
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '404':
                    $ref: '#/components/responses/404'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
    /cases/download:
        post:
            summary: >
                Streams a CSV, TSV or JSON attachment of cases matching the search query. 
                If caseIds are supplied the CSV only contains cases with those ids. 
                You cannot supply both query and caseIds.
            tags: [Case]
            operationId: downloadCases
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        schema:
                            type: object
                            properties:
                                caseIds:
                                    description: Cases corresponding to these ids will be returned
                                    type: array
                                    items:
                                        type: string
                                query:
                                    description: >
                                        Cases matching this query will be returned. This is ignored
                                        if empty.
                                    type: string
                                limit:
                                    description: >
                                        Max number of rows that can be downloaded when query is provided
                                    type: string
                                format:
                                    description: Format of data to be downloaded
                                    type: string
            responses:
                '200':
                    $ref: '#/components/responses/200CasesDownload'
                '500':
                    $ref: '#/components/responses/500'
    /cases/downloadAsync:
        post:
            summary: >
                Streams an attachment of cases matching the search query.
            tags: [Case]
            operationId: downloadAsync
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                query:
                                    description: >
                                        Cases matching this query will be returned.
                                    type: string
            responses:
                '200':
                    $ref: '#/components/responses/200FilteredDataSetDownload'
                '204':
                    $ref: '#/components/responses/204'
                '500':
                    $ref: '#/components/responses/500'
...
components:
    schemas:
...
        NewSource:
            description: >
                A "#/components/schemas/Source" with additional required fields for
                newly-created sources.
            allOf:
                - $ref: '#/components/schemas/Source'
            properties:
                origin:
                    required:
                        - url
                automation:
                    required:
                        - schedule
            required:
                - origin
...
    responses:
...
        '200SourceArray':
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/SourceArray'
...
        '204':
            description: Empty
        '400':
            description: Malformed request
...
        '403':
            description: Forbidden
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
...
        '422':
            description: Unprocessable entity
        '500':
            description: Internal server error
servers:
    - url: http://localhost:3001/api
      description: Local server
    - url: http://dev-data.covid-19.global.health/api
      description: Development server
    - url: http://data.covid-19.global.health/api
      description: Production server